split modules file to algorithm
