+++ neural_compressor/torch/algorithms/weight_only/__init__.py
@@ -15,3 +15,4 @@
 from .utility import *
 from .rtn import rtn_quantize
 from .gptq import gptq_quantize
+from modules import WeightOnlyLinear
