+++ test/3x/torch/quantization/weight_only/test_gptq.py
@@ -4,8 +4,8 @@ import pytest
 import torch
 import transformers
 
+from neural_compressor.torch.algorithms.weight_only import WeightOnlyLinear
 from neural_compressor.torch.quantization import GPTQConfig, get_default_gptq_config, get_default_rtn_config, quantize
-from neural_compressor.torch.quantization.modules import WeightOnlyLinear
 
 
 def run_fn(model):
