+++ test/3x/torch/quantization/weight_only/test_rtn.py
@@ -4,13 +4,13 @@ import pytest
 import torch
 import transformers
 
+from neural_compressor.torch.algorithms.weight_only import WeightOnlyLinear
 from neural_compressor.torch.quantization import (
     RTNConfig,
     get_default_double_quant_config,
     get_default_rtn_config,
     quantize,
 )
-from neural_compressor.torch.quantization.modules import WeightOnlyLinear
 
 
 class TestRTNQuant:
